Arduino sketch file
This file has the functionality for the portal gun.  Sound files need to be added to the Adafruit Sound f/x board locally.  Any sound file will work via drag-and-drop directly to the board, but Adafruit specifies the file type and the amount of memory for the sound files on their website.